WIP Chart to deploy this repo as near to form as possible
The point of this chart is to deploy the idiomatic https://github.com/tiangolo/full-stack-fastapi-template
as close to form as possible. This includes the use of traefik as an edge router, with entry points as defined
in the docker-compose files, routers defined in code, any middlewares (none currently), and supporting
services ad defined by containers which are expected to be deployed in additional helm chart(s). If this
helm chart were to be easily defined, it is to reproduce the exact same effect as the docker-compose command
but deployed to a k8s cluster, no more no less.